+++ changelog.md
@@ -13,6 +13,7 @@
 - Removed the now redundant [`eslint-plugin-import-order-alphabetical`](https://npm.im/eslint-plugin-import-order-alphabetical) dev dependency.
 - Stop using [`husky`](https://npm.im/husky) and [`lint-staged`](https://npm.im/lint-staged).
 - Use strict mode for scripts.
+- Fixed page `getInitialProps` not working when `withGraphQLApp` decorates an app that doesn’t have `getInitialProps`.
 
 ## 5.1.0